requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
h11==0.14.0
idna==3.7
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing-extensions==4.11.0
urllib3==2.2.1
wsproto==1.2.0
